Split AquaFlux AMM TVL into separate adapter
